useMemo, createRef and RefObject for menu
